KULRICE-5678 Third party dependency licensing is now embedded in all jar's at META-INF/THIRD-PARTY.txt.
Duplicate licenses are condensed under the correct license name.  Reduced licenses from 60+ down to 23.

Jenkins uses a plugin execution to fail a build job should a dependency be added to the build tree without a known license.

Process is in place to make this easy to keep up to date so it no longer requires a bunch of work at release time.

git-svn-id: https://svn.kuali.org/repos/rice/trunk@25136 7a7aa7f6-c479-11dc-97e2-85a2497f191d