Updated XPathQualifierResolver and ResolverConfig.xsd to allow "legacy" xml configurations to still be processed correctly. Changes allow xml config to be ingested where the resolverConfig on rule attributes is configured in either of the following formats:  ---------   <attributes name="standard">       <xPathExpression>./value</xPathExpression>    </attributes>  --- or ---   <qualifier name="legacy">       <xPathExpression>./value</xPathExpression>    </qualifier>   ---------
This is done because migration scripts were not made for the above change nor was it documented in impacting changes, and we like to maintain flexibility.

The test case was also updated, and commented appropriately. It now uses the configurations maintained as static strings and, for each test case, adds the standard config as well as a string replaced (qualifier for attributes) version, then runs the test for both values (everything in the set, actually). This should make it easy to add/remove if a similar change is made down the road and the test needs expanded or downsized.


Fixes KULRICE-7044 (Fixes https://jira.kuali.org/browse/KULRICE-7044)

(Also fixed a typo in the RuleAttribute.xsd that is inconsequential)

git-svn-id: https://svn.kuali.org/repos/rice/branches/rice-2.0@31977 7a7aa7f6-c479-11dc-97e2-85a2497f191d